many changes: - added 64-bit support (only q3a tested with ioq3) - combined windows and linux game_api tables into one - fixed ERROR log level not working - removed nocrash, greeting, and game log write features

+++ include/mod.h
@@ -23,7 +23,7 @@ typedef struct mod_s {
 	void* dll = nullptr;
 	qvm_t qvm = {};
 	mod_vmMain_t pfnvmMain = nullptr;
-	int vmbase = 0;
+	intptr_t vmbase = 0;
 	std::string path;
 } mod_t;
